fix(modal): move event up
