+++ src/components/modal/html/modal.tsx
@@ -1,65 +1,37 @@
-import { useEffect, useRef } from 'react';
-import type { MouseEventHandler } from 'react';
+import type { MouseEventHandler, ReactEventHandler, RefObject } from 'react';
 import type { LunaticComponentProps } from '../../type';
 import { createCustomizableLunaticField } from '../../commons';
 import { createPortal } from 'react-dom';
 import './modal.scss';
 import Button from '../../button';
-import { isEventInRect } from '../../../utils/dom';
 
 type Props = Pick<
 	LunaticComponentProps<'Modal'>,
-	| 'id'
-	| 'label'
-	| 'description'
-	| 'goToPage'
-	| 'page'
-	| 'goNextPage'
-	| 'goPreviousPage'
->;
+	'id' | 'label' | 'description'
+> & {
+	onClose: ReactEventHandler<HTMLDialogElement>;
+	onCancel: ReactEventHandler<HTMLDialogElement>;
+	onClick: MouseEventHandler<HTMLDialogElement>;
+	dialogRef: RefObject<HTMLDialogElement>;
+};
 
 function ModalPure({
 	id,
 	label,
 	description,
-	goNextPage,
-	goPreviousPage,
+	onClose,
+	onCancel,
+	onClick,
+	dialogRef,
 }: Props) {
-	const dialogRef = useRef<HTMLDialogElement>(null);
-
-	useEffect(() => {
-		// Browser behaviour is different for dialog displayed with this API, it creates de dialog with backdrop
-		dialogRef.current?.showModal();
-	}, []);
-
-	const handleClose = () => {
-		// A cancel button was pressed
-		if (dialogRef.current?.returnValue === 'cancel') {
-			handleCancel();
-			return;
-		}
-		goNextPage();
-	};
-
-	const handleCancel = () => {
-		goPreviousPage();
-	};
-
-	const handleDialogClick: MouseEventHandler<HTMLDialogElement> = (e) => {
-		// Cancel the modal on backdrop blick
-		if (!isEventInRect(e, dialogRef.current!.getBoundingClientRect())) {
-			handleCancel();
-		}
-	};
-
 	return createPortal(
 		<dialog
-			onClose={handleClose}
-			onCancel={handleCancel}
+			onClose={onClose}
+			onCancel={onCancel}
 			className="lunatic-modal"
 			ref={dialogRef}
 			id={id}
-			onClick={handleDialogClick}
+			onClick={onClick}
 		>
 			<form method="dialog">
 				<div className="lunatic-modal_title">{label}</div>
